High-Level Tuning interface added - documentation updated - examples updated, NXP MK64F1 example port added - lwIP NSD: core lock added - CLI callback signatures replaced with macros - CLI command removal macro added - fast compensation algorithm added

+++ src/flexptp/ptp_servo_types.h
@@ -20,7 +20,7 @@ typedef struct {
     // information about sync interval
     int8_t logMsgPeriod; ///< Logarithmic message period
     double msgPeriodMs; ///< Message period in ms
-    double measSyncPeriodNs; ///< Measured synchronization period (t1->t1)
+    int64_t measSyncPeriodNs; ///< Measured synchronization period (t1->t1)
 } PtpServoAuxInput;
 
 #endif /* FLEXPTP_SIM_PTP_SERVO_TYPES_H_ */
